Added reduction of the source model
